+++ Kyber/utils.h
@@ -3,7 +3,7 @@
 
 inline std::vector<uint8_t> bytesToBits(const uint8_t* b, int bLen) {
     std::vector<uint8_t> result(bLen * 8);
-    for (size_t i = 0; i < bLen; i++) {
+    for (int i = 0; i < bLen; i++) {
         for (int8_t j = 0; j < 8; j++) {
             result[8 * i + j] = (b[i] >> j) & 0x1;
         }
